edit commentModel and some authorization

+++ controller/commentController.js
@@ -2,10 +2,13 @@ const Comment = require('./../models/commentModel');
 const catchAsync = require('./../utils/catchAsync');
 const AppError = require('./../utils/appError');
 
-exports.setBookId = (req, res, next) => {
+exports.setBookUserIds = (req, res, next) => {
   // Allow nested routes
-  if (!req.body.book) req.body.book = req.params.bookId;
   // console.log(req.params.bookId);
+  // console.log(req.user.id);
+  if (!req.body.book) req.body.book = req.params.bookId;
+  if (!req.body.user) req.body.user = req.user.id;
+
   next();
 };
 
@@ -46,14 +49,24 @@ exports.createComment = catchAsync(async (req, res, next) => {
 
 exports.updateComment = catchAsync(async (req, res, next) => {
   const comment = await Comment.findById(req.params.id);
+
   if (!comment) {
-    return next(new AppError('No Comment found with that ID', 404));
+    return next(new AppError('No comment found with that ID', 404));
+  }
+  // Check if The Comment Belong To User
+  if (comment.user.toString() !== req.user.id.toString()) {
+    // console.log(comment.user.toString());
+    // console.log(req.user.id.toString());
+    return next(
+      new AppError('You are not allowed to update this comment', 403)
+    );
   }
 
   const doc = await Comment.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
     runValidators: true
   });
+
   res.status(200).json({
     status: 'success',
     data: {
@@ -64,8 +77,15 @@ exports.updateComment = catchAsync(async (req, res, next) => {
 
 exports.deleteComment = catchAsync(async (req, res, next) => {
   const comment = await Comment.findById(req.params.id);
+
   if (!comment) {
-    return next(new AppError('No Comment found with that ID', 404));
+    return next(new AppError('No comment found with that ID', 404));
+  }
+  // Check if The Comment Belong To User
+  if (comment.user.toString() !== req.user.id.toString()) {
+    return next(
+      new AppError('You are not allowed to delete this comment', 403)
+    );
   }
 
   await Comment.findByIdAndDelete(req.params.id);
